pyproject.toml
[project]
name = "loungeaccessadvisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore-starter-toolkit>=0.1.25",
    "strands-agents>=1.12.0",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.11",
    "streamlit>=1.50.0",
    "watchdog>=6.0.0",
    "boto3>=1.34.0",
    "bcrypt>=4.1.2",
]
